Move useful code from json_contains into a common place
Copied impl for mutable instance - which doesn't actually copy yet.
